execute middleware at correct time
